perbaikan 1

+++ app/Http/Controllers/KostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Kost;
+use App\Room;
 use App\Transformers\KostDetailTransformer;
 use App\Transformers\KostTransformer;
 use App\Transformers\RoomTransformer;
@@ -30,8 +31,9 @@ class KostController extends Controller
             ->toArray();
     }
 
-    public function create(Request $request, $id){
-        $user = User::find($id);
+    public function create(Request $request){
+        $user = Auth::user();
+
         if($user->role != 1)
             return response()->json([
                 'error'     => 'Tidak bisa menambahkan kost',
@@ -41,29 +43,37 @@ class KostController extends Controller
         $this->validate($request, [
             'nama'          => 'required',
             'deskripsi'     => 'required|min:5',
+            'address'       => 'required',
         ]);
 
         $kost = Kost::create([
-            'user_id'   => $id,
+            'user_id'   => $user->id,
             'nama'      => $request->nama,
             'deskripsi' => $request->deskripsi,
+            'address'   => $request->address,
         ]);
 
         return fractal($kost, new KostTransformer())->toArray();
     }
 
     public function delete($id){
+        $user = Auth::user();
         $kost = Kost::find($id);
-        if($kost == null)
-            return response()->json([
-                'error'     => 'Cant delete',
-                'message'   => 'Kost tidak ada',
-            ]);
 
+        // Cek kost apakah anda
+        if($kost == null)
+            return response()->json(['message' => 'kost tidak ada'], 400);
+
+        // Cek kepemilikan kost
+        if($user->id != $kost->user_id)
+            return response()->json(['message' => 'kost ini punya akun lain'], 403);
+
+        // Hapus semua roomnya dulu
+        Room::where('kost_id', $kost->id)->delete();
+
+        // Hapus kost
         $kost->delete();
-        return response()->json([
-            'message'   => 'Success',
-        ]);
+        return response()->json(['message' => 'kost berhasil dihapus']);
     }
 
     public function show(){
@@ -73,6 +83,15 @@ class KostController extends Controller
         return fractal()
             ->collection($kosts)
             ->transformWith(new KostTransformer())
+            ->toArray();
+    }
+
+    public function kostAndRoom(){
+        $user = Auth::user();
+
+        return fractal()
+            ->collection($user->kosts)
+            ->transformWith(new KostDetailTransformer())
             ->toArray();
     }
 }
